+++ contracts/Comptroller/Diamond/interfaces/IDiamondCut.sol
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
-pragma solidity 0.5.16;
-pragma experimental ABIEncoderV2;
+pragma solidity 0.8.25;
 
 interface IDiamondCut {
     enum FacetCutAction {