+++ contracts/Comptroller/Diamond/interfaces/IPolicyFacet.sol
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
-pragma solidity 0.5.16;
+pragma solidity 0.8.25;
 
 import { VToken } from "../../../Tokens/VTokens/VToken.sol";
 
@@ -95,4 +95,8 @@ interface IPolicyFacet {
         uint256[] calldata supplySpeeds,
         uint256[] calldata borrowSpeeds
     ) external;
+
+    function getBorrowingPower(
+        address account
+    ) external view returns (uint256 error, uint256 liquidity, uint256 shortfall);
 }