+++ contracts/Tokens/VAI/VAIControllerStorage.sol
@@ -1,4 +1,6 @@
-pragma solidity ^0.5.16;
+// SPDX-License-Identifier: BSD-3-Clause
+
+pragma solidity 0.8.25;
 
 import { ComptrollerInterface } from "../../Comptroller/ComptrollerInterface.sol";