+++ contracts/test/MockVBNB.sol
@@ -1,4 +1,4 @@
-pragma solidity ^0.5.16;
+pragma solidity 0.8.25;
 
 import "../Tokens/VTokens/VToken.sol";
 
@@ -20,15 +20,15 @@ contract MockVBNB is VToken {
      */
     constructor(
         ComptrollerInterface comptroller_,
-        InterestRateModel interestRateModel_,
+        InterestRateModelV8 interestRateModel_,
         uint initialExchangeRateMantissa_,
         string memory name_,
         string memory symbol_,
         uint8 decimals_,
         address payable admin_
-    ) public {
+    ) {
         // Creator of the contract is admin during initialization
-        admin = msg.sender;
+        admin = payable(msg.sender);
 
         initialize(comptroller_, interestRateModel_, initialExchangeRateMantissa_, name_, symbol_, decimals_);
 
@@ -39,7 +39,7 @@ contract MockVBNB is VToken {
     /**
      * @notice Send BNB to VBNB to mint
      */
-    function() external payable {
+    receive() external payable {
         (uint err, ) = mintInternal(msg.value);
         requireNoError(err, "mint failed");
     }
@@ -67,7 +67,7 @@ contract MockVBNB is VToken {
     // @custom:event Emits Transfer event on success
     // @custom:event Emits RedeemFee when fee is charged by the treasury
     function redeem(uint redeemTokens) external returns (uint) {
-        return redeemInternal(msg.sender, msg.sender, redeemTokens);
+        return redeemInternal(msg.sender, payable(msg.sender), redeemTokens);
     }
 
     /**
@@ -80,7 +80,7 @@ contract MockVBNB is VToken {
     // @custom:event Emits Transfer event on success
     // @custom:event Emits RedeemFee when fee is charged by the treasury
     function redeemUnderlying(uint redeemAmount) external returns (uint) {
-        return redeemUnderlyingInternal(msg.sender, msg.sender, redeemAmount);
+        return redeemUnderlyingInternal(msg.sender, payable(msg.sender), redeemAmount);
     }
 
     /**
@@ -90,7 +90,7 @@ contract MockVBNB is VToken {
      */
     // @custom:event Emits Borrow event on success
     function borrow(uint borrowAmount) external returns (uint) {
-        return borrowInternal(msg.sender, msg.sender, borrowAmount);
+        return borrowInternal(msg.sender, payable(msg.sender), borrowAmount);
     }
 
     /**
@@ -139,14 +139,14 @@ contract MockVBNB is VToken {
      * @param amount Amount of BNB being sent
      * @return The actual amount of BNB transferred
      */
-    function doTransferIn(address from, uint amount) internal returns (uint) {
+    function doTransferIn(address from, uint amount) internal override returns (uint) {
         // Sanity checks
         require(msg.sender == from, "sender mismatch");
         require(msg.value == amount, "value mismatch");
         return amount;
     }
 
-    function doTransferOut(address payable to, uint amount) internal {
+    function doTransferOut(address payable to, uint amount) internal override {
         /* Send the BNB, with minimal gas and revert on failure */
         to.transfer(amount);
     }
@@ -156,7 +156,7 @@ contract MockVBNB is VToken {
      * @dev This excludes the value of the current message, if any
      * @return The quantity of BNB owned by this contract
      */
-    function getCashPrior() internal view returns (uint) {
+    function getCashPrior() internal view override returns (uint) {
         (MathError err, uint startingBalance) = subUInt(address(this).balance, msg.value);
         require(err == MathError.NO_ERROR, "cash prior math error");
         return startingBalance;
@@ -197,7 +197,7 @@ contract MockVBNB is VToken {
      * @param reduceAmount Amount of reduction to reserves
      * @return uint Returns 0 on success, otherwise returns a failure code (see ErrorReporter.sol for details).
      */
-    function _reduceReservesFresh(uint reduceAmount) internal returns (uint) {
+    function _reduceReservesFresh(uint reduceAmount) internal override returns (uint) {
         // totalReserves - reduceAmount
         uint totalReservesNew;
 
@@ -244,7 +244,7 @@ contract MockVBNB is VToken {
      * @return uint Returns 0 on success, otherwise returns a failure code (see ErrorReporter.sol for details).
      */
     // @custom:event Emits ReservesReduced event
-    function _reduceReserves(uint reduceAmount) external nonReentrant returns (uint) {
+    function _reduceReserves(uint reduceAmount) external override nonReentrant returns (uint) {
         uint error = accrueInterest();
         if (error != uint(Error.NO_ERROR)) {
             // accrueInterest emits logs on errors, but on top of that we want to log the fact that an attempted reduce reserves failed.
@@ -260,7 +260,7 @@ contract MockVBNB is VToken {
      *   up to the current block and writes new checkpoint to storage.
      */
     // @custom:event Emits AccrueInterest event
-    function accrueInterest() public returns (uint) {
+    function accrueInterest() public override returns (uint) {
         /* Remember the initial block number */
         uint currentBlockNumber = getBlockNumber();
         uint accrualBlockNumberPrior = accrualBlockNumber;