+++ helpers/markets/bscmainnet.ts
@@ -120,8 +120,8 @@ export default [
     asset: "BTCB",
     interestRateModel: {
       model: "jump",
-      baseRatePerYear: "0",
-      multiplierPerYear: "0.09",
+      baseRatePerYear: "0.0025",
+      multiplierPerYear: "0.0367",
       kink: "0.75",
       jumpMultiplierPerYear: "2",
     },
@@ -751,6 +751,31 @@ export default [
     },
     initialSupply: {
       amount: "0.14",
+      vTokenReceiver: "VTreasury",
+    },
+  },
+  {
+    name: "Venus WBNB",
+    symbol: "vWBNB",
+    asset: "WBNB",
+    interestRateModel: {
+      model: "two-kinks",
+      baseRatePerYear: "0",
+      multiplierPerYear: "0.045",
+      kink: "0.7",
+      baseRatePerYear2: "0",
+      multiplierPerYear2: "1.4",
+      kink2: "0.8",
+      jumpMultiplierPerYear: "3",
+    },
+    riskParameters: {
+      collateralFactor: "0.8",
+      reserveFactor: "0.3",
+      supplyCap: "2672000",
+      borrowCap: "2008000",
+    },
+    initialSupply: {
+      amount: "1",
       vTokenReceiver: "VTreasury",
     },
   },